Small authentication cleanups
