+++ apps/app/src/styles/auth.ts
@@ -2,6 +2,7 @@ import { StyleSheet } from 'react-native';
 
 const styles = StyleSheet.create({
 	container: {
+		flex: 1,
 		justifyContent: 'center',
 		padding: 20
 	}
